pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "windsurf-dreambooth"
version = "0.1.0"
description = "DreamBooth Studio - A modular application for fine-tuning and generating images with Diffusion models"
dependencies = [
    "torch",
    "torchvision",
    "setuptools",
    "fastapi",
    "uvicorn[standard]",
    "diffusers",
    "transformers",
    "accelerate",
    "python-multipart",
    "pillow",
    "numpy",
    "gradio",
    "huggingface-hub",
    "huggingface_hub[hf_xet]",
    "loguru",
    "sentencepiece",
    "protobuf"
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

# Exclude non-package directories to avoid multiple-package discovery
[tool.setuptools.packages.find]
exclude = ["outputs", "outputs.*", "frontend", "frontend.*", "finetuned_models", "finetuned_models.*"]